cvOk() - Push command up from PathCommandTest to CivilTestCase
