+++ tests/CivilTestCase.php
@@ -3,6 +3,7 @@ namespace Civi\Cv;
 
 use Civi\Cv\Util\Process as ProcessUtil;
 use Civi\Cv\Application;
+use Civi\Cv\Util\Process;
 use Symfony\Component\Console\Tester\CommandTester;
 use Symfony\Component\Filesystem\Filesystem;
 
@@ -39,7 +40,11 @@ class CivilTestCase extends \PHPUnit_Framework_TestCase {
   }
 
   /**
+   * Create a process for `cv` subcommand.
+   *
    * @param string $command
+   *   Ex: "vars:show --out=serialize".
+   * @return \Symfony\Component\Process\Process
    */
   protected function cv($command) {
     $process = new \Symfony\Component\Process\Process("{$this->cv} $command");
@@ -61,6 +66,39 @@ class CivilTestCase extends \PHPUnit_Framework_TestCase {
     $commandTester = new CommandTester($command);
     $commandTester->execute($args);
     return $commandTester;
+  }
+
+  /**
+   * Run a `cv` subcommand. Assert success and return output.
+   *
+   * @param string $cmd
+   * @return string
+   */
+  protected function cvOk($cmd) {
+    $p = Process::runOk($this->cv($cmd));
+    return $p->getOutput();
+  }
+
+  /**
+   * Run a `cv` subcommand. Assert success and return output.
+   *
+   * @param string $cmd
+   * @return string
+   */
+  protected function cvFail($cmd) {
+    $p = Process::runFail($this->cv($cmd));
+    return $p->getErrorOutput() . $p->getOutput();
+  }
+
+  /**
+   * Run a `cv` subcommand. Assert success and return output.
+   *
+   * @param string $cmd
+   * @return string
+   */
+  protected function cvJsonOk($cmd) {
+    $p = Process::runOk($this->cv($cmd));
+    return json_decode($p->getOutput(), 1);
   }
 
 }
